requirements.txt
pdfplumber
python-docx
sentence-transformers
numpy
torch
ctransformers
